+++ markdown.d.ts
@@ -785,6 +785,7 @@ export namespace MarkdownParser {
         SourceParser.StrParser,
         AutolinkParser.HashtagParser,
         AutolinkParser.HashrefParser,
+        SourceParser.StrParser,
       ], Context> {
     }
     export namespace AutolinkParser {