+++ src/parser/inline/autolink/hashtag.ts
@@ -3,5 +3,5 @@ import { validate, focus, creator } from '../../../combinator';
 import { html } from 'typed-dom';
 
 export const hashtag: AutolinkParser.HashtagParser = creator(validate('#', focus(
-  /^#(?![0-9]+(?![A-Za-z]|[^\x00-\x7F\s]))(?:[A-Za-z0-9]|[^\x00-\x7F\s])+/,
+  /^#(?![0-9])(?:[A-Za-z0-9]|[^\x00-\x7F\s])+/,
   tag => [[html('a', { class: 'hashtag', rel: 'noopener' }, tag)], ''])));