+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.151.0
 
 - Extend account syntax to be possible to clarify the namespace.
+- Change hashtag syntax to disallow strings starting with a number.
 - Change reference syntax to enable extension syntax.
 
 ## 0.150.0